Update intro.md
Feel free to ignore this one, as this might be seen as not appropriate. However, to a new user of MO2, what seem to an experienced user a "normal procedural message" to an experienced user, seem like an error message to a newbie, and would be better adressed before the mods than after in a faq.

I've copy-pasted the faq section to the beginning and this might not be the most appropriate way to prepare a user but since this is the suggested way to make suggestions, I've done it like that.