Update finish.md
I don't know exactly what change would be good here to address the confusion that might result from the player not having this particular "spicy" mod, or really, any other mod list resulting from the player choses of mod to install. Well, not sure exactly how to do it without ruining the guide and make it more complicated for nothing.

So for now I just suggested to change "should" to "could".